Replace nodejs make.js with a make.js for deno
